Move unix and udp networking files into separate directories
