+++ src/software/BUILD
@@ -17,9 +17,9 @@ cc_binary(
         "//software/estop:arduino_util",
         "//software/logger",
         "//software/multithreading:observer_subject_adapter",
-        "//software/networking:threaded_proto_udp_listener",
-        "//software/networking:threaded_proto_udp_sender",
-        "//software/networking:threaded_proto_unix_listener",
+        "//software/networking/udp:threaded_proto_udp_listener",
+        "//software/networking/udp:threaded_proto_udp_sender",
+        "//software/networking/unix:threaded_proto_unix_listener",
         "//software/sensor_fusion:threaded_sensor_fusion",
         "//software/util/generic_factory",
         "@boost//:program_options",
@@ -31,8 +31,8 @@ cc_binary(
     srcs = ["er_force_simulator_main.cpp"],
     deps = [
         "//software:constants",
-        "//software/networking:threaded_proto_unix_listener",
-        "//software/networking:threaded_proto_unix_sender",
+        "//software/networking/unix:threaded_proto_unix_listener",
+        "//software/networking/unix:threaded_proto_unix_sender",
         "//software/simulation:er_force_simulator",
         "@boost//:program_options",
     ],
@@ -44,7 +44,7 @@ cc_binary(
     deps = [
         "//proto:tbots_cc_proto",
         "//shared:constants",
-        "//software/networking:threaded_proto_udp_listener",
+        "//software/networking/udp:threaded_proto_udp_listener",
         "@boost//:program_options",
     ],
 )
@@ -88,8 +88,8 @@ pybind_library(
         "//software/geom:segment",
         "//software/geom:vector",
         "//software/geom/algorithms",
-        "//software/networking:threaded_proto_udp_listener",
-        "//software/networking:threaded_proto_udp_sender",
+        "//software/networking/udp:threaded_proto_udp_listener",
+        "//software/networking/udp:threaded_proto_udp_sender",
         "//software/uart:boost_uart_communication",
         "//software/world",
         "//software/world:field",
