+++ src/software/jetson_nano/services/network/BUILD
@@ -9,8 +9,8 @@ cc_library(
         "//proto:tbots_cc_proto",
         "//shared:robot_constants",
         "//software/logger",
-        "//software/networking:threaded_proto_udp_listener",
-        "//software/networking:threaded_proto_udp_sender",
+        "//software/networking/udp:threaded_proto_udp_listener",
+        "//software/networking/udp:threaded_proto_udp_sender",
         "@boost//:asio",
     ],
 )
