+++ src/software/jetson_nano/services/network/network.h
@@ -8,8 +8,8 @@
 #include "shared/constants.h"
 #include "shared/robot_constants.h"
 #include "software/jetson_nano/services/network/proto_tracker.h"
-#include "software/networking/threaded_proto_udp_listener.hpp"
-#include "software/networking/threaded_proto_udp_sender.hpp"
+#include "software/networking/udp/threaded_proto_udp_listener.hpp"
+#include "software/networking/udp/threaded_proto_udp_sender.hpp"
 
 class NetworkService
 {
