+++ src/software/logger/BUILD
@@ -117,7 +117,7 @@ cc_library(
         "//proto/message_translation:tbots_protobuf",
         "//shared:constants",
         "//software/logger:plotjuggler_sink",
-        "//software/networking:threaded_proto_udp_sender",
+        "//software/networking/udp:threaded_proto_udp_sender",
         "@base64",
         "@g3log",
     ],
@@ -135,7 +135,7 @@ cc_library(
     deps = [
         "//proto:visualization_cc_proto",
         "//shared:constants",
-        "//software/networking:threaded_udp_sender",
+        "//software/networking/udp:threaded_udp_sender",
         "@g3log",
     ],
 )
@@ -153,7 +153,7 @@ cc_library(
         "//proto:any_cc_proto",
         "//proto:tbots_cc_proto",
         "//shared:constants",
-        "//software/networking:threaded_unix_sender",
+        "//software/networking/unix:threaded_unix_sender",
         "@base64",
         "@g3log",
     ],
