+++ src/software/logger/network_sink.h
@@ -6,7 +6,7 @@
 #include "proto/robot_log_msg.pb.h"
 #include "proto/visualization.pb.h"
 #include "software/logger/log_merger.h"
-#include "software/networking/threaded_proto_udp_sender.hpp"
+#include "software/networking/udp/threaded_proto_udp_sender.hpp"
 
 
 /**
