+++ src/software/networking/unix/threaded_unix_sender.h
@@ -4,7 +4,7 @@
 #include <boost/bind.hpp>
 #include <string>
 
-#include "software/networking/unix_sender.h"
+#include "software/networking/unix/unix_sender.h"
 
 class ThreadedUnixSender
 {
