+++ src/software/simulated_tests/BUILD
@@ -66,8 +66,8 @@ py_library(
         ":tbots_test_runner",
         "//proto:import_all_protos",
         "//software/logger:py_logger",
-        "//software/networking:threaded_unix_listener_py",
-        "//software/networking:threaded_unix_sender_py",
+        "//software/networking/unix:threaded_unix_listener_py",
+        "//software/networking/unix:threaded_unix_sender_py",
         "//software/simulated_tests:validation",
         "//software/thunderscope",
         "//software/thunderscope/binary_context_managers:full_system",
@@ -88,8 +88,8 @@ py_library(
     deps = [
         "//proto:import_all_protos",
         "//software/logger:py_logger",
-        "//software/networking:threaded_unix_listener_py",
-        "//software/networking:threaded_unix_sender_py",
+        "//software/networking/unix:threaded_unix_listener_py",
+        "//software/networking/unix:threaded_unix_sender_py",
         "//software/simulated_tests:validation",
         "//software/thunderscope",
         "//software/thunderscope:config",
